[api] Add add_enrollment() function
The function adds a new relationship between the unique identity
identified by 'uuid' and the given 'organization'. The given
identity and organization must exist prior to add this enrollment.
The period of the enrollment can be given with the parameters 'start_date'
and 'end_date'.

If the given enrollment data is already in the registry, the function
will raise an exception.