[api] Support top domains
With 'is_top_domain' flag, a domain can be set as a top domain.
This is useful to avoid the insertion of sub-domains which belong
to the same organization (i.e  eu.example.com, us.example.com).
This flag can be updated using 'overwrite' parameter.

Model and tests were also updated to support this new feature.