[api] Check the existence of a UUID when a new identity is created
So far, add_identity() function only checked whether a tuple of
source, name, email and username existed on the database before
the insertion of a new identity. With this patch, the UUID is also
checked to avoid problems when tuples with the combination of
"None" and None values are inserted.