[api] Check identity id uniqueness in add_identity()
So far, sortinghat did not check whether the id of an identity
was already inserted in the database. It only checked if the
tuple (source, name, email, username) was stored. This might
raise some problems like in #67, when the values of an identity
are truncated because their length, so they were not found in the
database when they are compared.

With this patch this problem will not happen again because
for those cases, the UUID will be the same and a AlreadyExistError
will be raised. If for any reason the UUIDs are different,
it still have a unique constraint in the identities table that
checks if new tuples of (source, name, email, username) to be inserted
already exist in the database.

The problem with the length of the values will persist and
they will be truncated.