[database] Fix TIME_WAIT connections error
For unknown reasons, when an exception is raised when a
database connection is in use, the connection will not close
properly, leaving it in a TIME_WAIT state. If this happens
many times, for example when a huge amount of identities are
affiliated, MySQL servers ignore new incoming connections
util those open are closed by the kernel.

This behaviour only occurs using NullPool class. That class
closes connections, when a session is over. Replacing it by
QueuePool class fixes this issue because it reuses connections.

The use of this class raises the "Too many connections" error
when test cases are run (see 693837f for further info). This
commit also fixes the unit tests reusing the connections.