[utils] Use case insensitive seed to generate UUIDs
So far, the identities UUIDs are generated using case sensitive
values with the seed (source:email:name:username). This means
that for an identity with the same values in upper or lower case
(i.e: jsmith@example.com and JSMITH@example.com) different UUIDs
are generated when the identity is the same.

This patch fixes this problem generating UUIDs ignoring the case
of the values.

Take into account that it produces an incompatibility with
previous UUIDs. These UUIDs must be regenerated.