[utils] Use unaccent name values to generate UUIDs
So far, the identities UUIDs are generated using accent characters
This means that for an identity with the same values accent or
unaccent (i.e: John Smith or Jöhn Smith), different UUIDs
are generated when the identity is the same.

This patch fixes this problem generating UUIDs converting this
accent character to their canonical form.

This change is caused by the behaviour of MySQL with unaccent
or accent characters. MySQL considers those characters the same,
raising IntegrityError exceptions when similar tuples with accent
or unaccent values are inserted into the database. Generating
the same UUID for these cases will prevent the error.

Take into account that it produces an incompatibility with
previous UUIDs. These UUIDs must be regenerated.