[model] Force 'utf8_unicode_ci' collation on MySQL tables
MySQL considers chars like 'β' and 'b' or 'ı' and 'i' the same,
when some collation values are set (i.e utf8_general_ci). This can
raise integrity errors when Sorting Hat tries to add similar
identities with these pairs of characters.

For instance, if the identity:

    ('scm', 'βart', 'bart@example.com', 'bart)

is stored in the database, the insertion of:

    ('scm', 'bart', 'bart@example.com', 'bart)

will raise an error, even when these identities have different
UUIDs.

Forcing MySQL to use 'utf8_unicode_ci' fixes this error, allowing
to insert both identities.