[model][tests] Define an order for domains and adapt tests to MariaDB
In models.py, add an order for domains, so that they are
ordered with respect to organizations.

Adapt testing to these changes.

In addition, adapt to testing using MariaDB as backend.
MariaDB has some peculiarities when using it from SQLAlchemy.
These changes address them. Most of the changes are for the
tests themselves, except for the one to database.py, which
includes a new exception which is raised by MariaDB.
Now, all the tests pass with MariaDB.

This caused some pain for tests with MySQL, because some
stuff was ordered differently now. So, in this patch there
are some minot tweaks to the tests to ensure the order
in some cases is the same for both database (and, at the
same time, to avoid randomness). Now, all tests pass
with MySQL too.

In the process, a bug in utils.merge_date_ranges() was
uncovered, and fixed.