[db] Fix tables created with invalid collation
In some random situations SortingHat tables appear with an invalid
collation. This is related to a wrong generation of the DDL table
statement by SqlAlchemy, which may randomly prepend the collation
information (mysql_collate) to the charset one (mysql_charset),
causing the former to be ignored. Changing mysql_charset to
mysql_default_charset fixes the problem.