[tests] Create tests database before running tests.
I got a bit fed up of having to create the database for tests
before actually running the tests. And anyway, for automation
it is nice that you just "run_tests.py". So, I instrumented
the base class for all database tests, so that it creates
(only once in a test run) the database before running all the
tests.

Since now it is not needed, I also removed the dabase creation
line from the travis config.