[autoprofile] Avoid exception when no id in selected data sources
Fixes #122

In some cases, a uuid has only ids for data sources which are not
among those selected for autoprofiling. In this case, it may happen
that one of those uuids is selected for autoprofiling, and no
id is going to be found (because only those corresponding to
the selected data sources are considered). This patch catches
the exception that is raised in that case.

A better fix is described in #122.