[unify] Add resume method to unify command
This code proposes a method to recover failed unify executions.
In a nutshell, after calculating the matches in memory, they are
converted to JSON items and marked as not processed. Each match
successfully executed is marked as processed. In case of a failure,
the failed match plus the remaining ones are serialized to disk
(which location is fixed). In the next execution, those matches
will be loaded and re-executed again, if the recovery option is
enabled.