[tests] Adapt tests to run with setup.py test
In the migration to get all of GrimoireLab modules to run the
tests via setup.py, some stuff waas needed, which is in this
patch:

* Adaption of setup.py itself, including testing dependencies,
adapting some other details in the script.
* Adaption of tests to be run from the parent directory.
* Create the database to use for testing, and remove it after the
tests.
* New option in tests.conf, for running that code to
create the database or not.
* Adapt the README.md to show how tests are run.
* Adapt .travis.yml to run with setup.py (and update it a bit).

In the process for adapting the tests to be run from the parent
directory, their portability was increased: now they should run
also in systems where "/" is not the path separator.