[db] Handle disconnection using pessimistic mode
SQLAlchemy offers a pessimistic mode to handle database
disconnection. Setting 'pool_pre_ping' parameter on the
database engine will check if the database connection
is still active when a session of the connection pool
is reused. To check that it runs a simple SQL statement.
This causes a small hit in the performance but it's worth
it.

This feature was included in 1.2 version, so the list
of dependencies has been updated as well.