[db] Support UTF-8 fully in MySQL/MariaDB
The default charset of UTF-8 (utf8) in MySQL/MariaDB does
not support, even when they are part of the standard,
4-bytes long characters. This means characters like emojis
or certain chinese characters cannot be inserted. Usually,
identities names or usernames have these types of characters.

The charset that fully supports UTF-8 is 'utf8mb4' using the
collation 'utf8mb4_unicode_520_ci'. This collation implements
the suggested Unicode Collation Algorithm (v5.2).

Using 'utf8mb4' also implies that the maximum size of char
(VARCHAR and so on) columns is 191. Indexes cannot be larger
than that when using InnoDB engine.

This patch can break the database so it is recommended to
regenerate again the database from a dump.

More info:
  * https://dev.mysql.com/worklog/task/?id=2673
  * https://dev.mysql.com/doc/refman/5.5/en/charset-unicode-conversion.html
  * https://mathiasbynens.be/notes/mysql-utf8mb4