[model] Migrate model to Django ORM
The previous Sorting Hat model was based on SQLAlchemy ORM.
This one preserves the same structure but adds two new fields
to each object. These two new fields are 'created_at' and
'last_modified'.

The first field will be set only to the date when an object
is created. The second one will be set automatically when an
object is modified.

Along with this commit, it's been included the migrations file
used by Django to setup the database; the requirements file
with the Python packages needed to run this software and
a directory with unit tests.