[graphql] Define basic GraphQL query to fetch organizations
This commit defines a basic schema and a query to retrieve
the information stored by an organization. To ask for the
list of organizations and their domains stored by SortingHat,
a GraphQL query would be like this:

```
{
  organizations {
    name
    domains {
      domain
      isTopDomain
    }
  }
}
```

GraphQL support is provided by `graphene` and `graphene_django`
libraries. The project configuration has been modified to
set the GraphQL web client for testing purposes. This client
is accessible through '/graphql' URL. In a future, it will be
removed from the standard configuration.