[graphql] Define GraphQL mutation to add domains
The mutation 'addDomain' allows to add a domain to the
registry. Errors are returned in case the organization
where this domain added does not exist, when the domain
name is empty or when it already exists in the registry.

To add a domain a mutation query like this will work:

  mutation addDom {
    addDomain(organization: "Example", domain: "example.net") {
      domain {
        domain
        isTopDomain
        organization {
          name
        }
      }
    }
  }