[graphql] Define GraphQL mutation to add identities
The mutation 'addIdentity' allows to add a new identity
to the registry. These identities can be added to
an existing unique identity. By defualt, a new unique
identity will be created which each new identity.

Errors are returned in case the identity already exists,
when the unique identity to link the new identity
does not exist or when any parameter has an invalid value.

To add an identity a mutation query like this will work:

mutation addId(
  $source: String,
  $name: String,
  $email: String,
  $username: String
  $uuid: String) {
  addIdentity(
    source: $source
    name: $name
    email: $email
    username: $username
    uuid: $uuid) {
      uuid
      identity {
        id
        name
        email
        username
        source
      }
    }
  }