[graphql] Define GraphQL mutation to delete identities
The mutation 'deleteIdentity' allows to delete existing
identities or unique identities from the registry.
By default, it removes identities but when the given
UUID belongs to a unique identity, that will be deleted.

Errors are returned either when the identity does not exist,
or when the UUID has an invalid value.

To remove an identity a mutation query like this will work:

mutation delId($uuid: String) {
  deleteIdentity(uuid: $uuid) {
    uuid
    uidentity {
      uuid
      identities {
        id
        name
        email
        username
        source
      }
    }
  }
}