[graphql] Define GraphQL mutation to update profiles
The mutation 'updateProfile' allows to update unique
identities profiles. Many attributes can be set such
as the name or email, the gender and whether the unique
identity is a bot or not.

To update the profile through a mutation, a query like
this should work:

mutation editProfile($uuid: String, $data: ProfileInputType) {
  updateProfile(uuid: $uuid, data: $data) {
    uuid
    uidentity {
      uuid
      profile {
        name
        email
        gender
        genderAcc
        isBot
        country {
          name
          code
        }
      }
    }
  }
}