[graphql] Allow filtering uidentities by uuid
In order to get one uidentity by its uuid we added a filter to the
query. Moreover, a new type for filtering has been created in
order to add filters easily without changing the query, so we can
filter in more than one field using this new type.

This query expects these three behaviors:

- If you specify the uuid in the filter you will receive the
  uidentites that match that uuid.
- If you don't put the uuid filter, you will receive all the
  uidentites.
- If you put an uuid in the filter that does not exist,
  you will receive an empty list.

There are also tests in order to check that it works as expected.