[parsing:grimoirelab] Add support to parse a blacklist
While SortingHat standard file includes a list of
blacklisted entries, GrimoireLab format does not.

This patch allows to include a blacklist on the YAML
file, that will be parsed and converted into the
SortingHat format.

Take into account the inclusion of a blacklist is
totally optional.

This patch also updates `grimoirelab2sh` script
to support this new feature.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>