[setup] Set pandas dep version to 0.22.0
This code changes the version of the Pandas dependency
to 0.22.0. This change is needed since in Travis the
max version distribution for Pandas is currently
0.22.0. No problems appear when executing the Travis CI
for sortinghat with a higher version (0.24.2), since
the latter version is downloaded and compiled. This
doesn't happen when executing Travis tests
for components that have a dependency with
SortingHat (i.e., ELK and Mordred). In this case,
Travis relies on the version distributions available
(<= 0.22.0).

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>