[db] Validate organization name field when performing a search
Following the checks which are performed in other methods
from `db.py`, now the method `find_organization` validates that
the organization name used for performing a search is not an
empty string, is not only composed by whitespaces and it is not
`None`.

The tests for the `add_organization` and `delete_organization`
methods from `api.py` have been extended checking values that
can cause this error, as they call `find_organization` method
internally.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>