[api] Add transactions in all API methods calling DB methods
Now, a `TransactionsLog` object is created for every method
calling a DB method. This object is passed as argument and
will be part of a set of operations created by DB methods.

Some existing unit tests were modified to check if no additional
transactions are created after an error occurs; and there are new
unit tests checking that every API method is generating the
transactions and operations that are expected.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>