[schema] Add queries for `Transactions` and `Operations`
These new queries allow to export the list of transactions
and operations from the database with the following format:

query {
    transactions {
      name
      createdAt
      tuid
      isClosed
      closedAt
    }
}

query {
    operations {
      ouid
      opType
      entityType
      target
      timestamp
      args
      trx {
        name
        createdAt
      }
    }
}

Both queries support filtering for most of the fields from
their corresponding models, including `from_date` and `to_date`
filters for both models (referencing to `created_at` field
for transactions and `timestamp` for operations). This date
can be passed as filter as a string following ISO 8601 format.

As the `args` field from `Operations` model is encoded in a
serialized JSON, it was necessary to define a explicit conversion
to return these encoded JSON fields as dictionaries. The class
in charge of this process is named `OperationArgsType` and it
inherits from `GenericScalar` class, which is a basic type defined
in `graphene` module representing a generic GraphQL scalar value
that could be String, Boolean, Int, Float, List or Object.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>