[schema] Add support for pagination to exisiting queries
Pagination for query results has been implemented on top of
the `Paginator` object from Django and applied to the existing
queries.

When performing a query, there are two new parameters:
- `pageSize`: Number of elements we want to obtain per page.
 If `pageSize` value is not set, the value will be set using
 a configuration parameter defined in Django settings, named
 `DEFAULT_GRAPHQL_PAGE_SIZE`.
- `page`: Number of the page we are asking for.
 If `page` value is not set, it will return the first page
 by default.

These changes modify the structure of the queries. Now, the
returned objects will be given under `entities`, and the
information related with the pagination under `pageInfo`.
For instance, to ask for a list of unique identities using
pagination, the query would look like this:

```
{
  uidentities(
    page: 2
    pageSize: 5
  ){
    entities{
      uuid
    }
    pageInfo{
      page
      pageSize
      numPages
      hasNext
      hasPrev
      startIndex
      endIndex
      totalResults
    }
  }
}
```

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>