[database] Allow SSL connections
The engine was not configure to accept SSL connections.
With this patch, the database module will try to establish
a SSL connection first. If this is not possible, it will
try to communicate with the server with a non-secure
connection.

To set the SSL connection, it was necessary a minor hack
that consists on passing an non-empty dictionary with key
'ssl' in 'connect_args'. The module pymysql expects a dict
with some SSL paramters, which will force to use specific
cert o ca. As we do not want that, we found that passing
a dict with a random key is valid and pymsql ignores that
key but enforces a SSL connection.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>