[setup] Let pandas be >= 0.22.0
If we pin pandas at 0.22.0, it doesn't build with Python 3.7.x,
apparently because of problems when building numpy for pandas.
Letting pandas be greater or equal than 0.22.0 doesn't seem
to break anything, since platforms having only 0.22.0,
as Travis had in July'19 (see #206), would still fulfill the
dependency.

I also set it to be less or equal than 0.25.3 because 1.0.0rc0
is out, which Travis is pulling when building for python 3.5,
but apparently it includes the new formating strings which
seem not to be supported in python 3.5.

Signed-off-by: Jesus M. Gonzalez-Barahona <jgb@gsyc.es>