[config] Hide Tracebacks caused by controlled Graphene errors
After this change, the logger object from Graphene has been
set to show only Fatal/Critical errors.

In Graphene-Django module, all the exceptions raised which are
caused by a GraphQL execution returns a result detailing the
errors. However, even if the exception is properly captured,
a Traceback is printed to STDERR where the Django `manage.py`
server is executed.

This was producing undesired error messages whenever a query
caused a controlled error like, for instance, adding a
duplicated entry to the database.

There is not an official solution for this problem yet, so we
decided to go for this workaround, as the Traceback is not
displayed anymore but the query result is still including the
information from results.errors, including what caused the
error or exception when executing the query.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>