[schema] Add support for authenticated queries
These changes add support for performing authenticated queries
(both GraphQL queries and mutations), based on JSON Web Tokens
(JWT).

To perform an authenticated query, an existing user must
generate a token which has to be included in the `Authorization`
header with the HTTP request. This token is generated using a
mutation which comes defined by the `graphene-jwt` module,
together with other extra mutations for verifying an existing
token and refreshing it in case the token expired.

By default, all the queries and mutations now require
authentication. This is controlled by a custom decorator named
`check_auth` which is based on the pre-defined decorator
`login_required` and extending it to check both if the `user`
object from the `context.info` object is authenticated and the
value of a new control variable named `AUTHENTICATION_REQUIRED`,
defined in Django settings allowing to set whether authentication
is required or not.

If the user is not logged or uses an invalid token, the following
exception is raised:

```
graphql.error.located_error.GraphQLLocatedError: You do not have
permission to perform this action
```

Note: After these changes, the `AUTHENTICATION_REQUIRED` variable
in the testing configuration script must be set to `True`.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>