[api] Add support for merging multiple identities at once
Before this change, the `merge_identities` method accepted
two unique identities to be merged: `from_uuid` and `to_uuid`.

After this change, the method accepts a list of unique identities
`from_uuids` to be merged at once into `to_uuid`.

This method has been refactored by including some of the logic
as inner methods, and the existing inner methods have been adapted
for handling a list of identities instead of a single one.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>