[log] Add a context as new input value for TransactionsLog class
With this change, all `TransactionsLog` objects have to be
instanciated with one more field containing the context where the
transaction is created.

This context is a `namedtuple` called `SortingHatContext`, which
contains a field named `user`. This user field must be an instance
of a Django `User` or `AnonymousUser` classes.

From this `user` field inside the context, the value of the field
`username` is used for setting the value of `authored_by` field
from the `Transaction` model.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>