[api] Fill profile information when a unique identity is created
Before this change, when a new unique identity was added by using
the API method `add_identity`, it may receive some optional input
parameters such as the name, the email and the source. This
information was set to the main sub-identity, but it was not
propagated to the unique identity's profile, which remained empty.

After this change, the profile fields `name` and `email` are set
with the corresponding input values from `add_identity` method.

This change affects both to `add_identity` and `move_identity` API
methods, which are the ones where new unique identities are
created.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>