[db-model] Remove CoerceToBool type
This code fixes the bug #121, thus it removes the type
CoerceToBool (introduced at 5f69ed8), which was used to handle
string representations of boolean values (i.e., True and False)
for SqlAlchemy versions prior to 1.1.14. This was due to the fact
that SqlAlchemy versions 1.1.x defined the Boolean type as follows:
```
Boolean typically uses BOOLEAN or SMALLINT on the DDL side,
and on the Python side deals in True or False.
```

Thus, equivalent string representations weren't handled
and required the definition of the CoerceToBool (which created an
additional CHECK constraint).

From versions greater than >=1.2.0 (already set in the current setup.py),
the CoerceToBool isn't needed anymore and can be removed. In particular, the
Boolean type is now defined as:
```
The Boolean datatype currently has two levels of assertion that the values
persisted are simple true/false values. For all backends, only the Python
values None, True, False, 1 or 0 are accepted as parameter values. For those
backends that don’t support a “native boolean” datatype, a CHECK constraint
is also created on the target column. Production of the CHECK constraint
can be disabled by passing the Boolean.create_constraint flag set to False
```

Tests have been updated accordingly.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>