[api] Add method for unmerging identities
Similar to `merge_identities`, this method accepts a list of
identities as input, so multiple identities can be
unmerged/splitted at once, creating a unique identity for
each one.

`name` and `email` fields from every identity are set to the
corresponding profile fields when the new unique identity is
created. Nor the enrollments or the profile from any parent
unique identity of the input identities are modified.

When a given identity `uuid` is equal to the `uuid` from its
parent unique identity, there will be no effect.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>