[schema] Add mutation for unmerging identities
Similar to `merge_identities`, this mutation accepts a list
of `uuid`s as input, allowing to unmerge several identities
from their parent unique identity at once, even if it is not
the same for all of them.

If the `uuid` from the parent unique identity is included in
the input list, this will have no effect.

Mutation example:

```
mutation unmergeIds($uuids: [String]) {
  unmergeIdentities(uuids: $uuids) {
    uuids
    uidentities {
      uuid
      identities {
        id
        name
        email
        username
        source
      }
    }
  }
}
```

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>