[schema] Add mutations for locking and unlocking unique identities
The input value for both `lock` and `unlock` mutations is the
`uuid` of the unique identity to be locked or unlocked.

Note that these actions also affect to the unique identity's
related objects such as identities, enrollments or its profile.

This `isLocked` new field has been added as a new filter for
the query asking for the unique identities.

Mutation examples:

"""
mutation lockId($uuid: String) {
            lockIdentity(uuid: $uuid) {
              uuid
              uidentity {
                uuid
                isLocked
              }
            }
          }
"""
"""
mutation unlockId($uuid: String) {
            unlockIdentity(uuid: $uuid) {
              uuid
              uidentity {
                uuid
                isLocked
              }
            }
          }
"""

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>