[api] Fix move when from identity exists as unique identity
Before this change, two problematic cases were found:
* When `from_id` was the only identity of a UniqueIdentity,
 move operation made the unique identity got orphan.
* When `from_id` matched the `uuid` of its parent
 UniqueIndentity, the sub-identity was moved to a different
 UniqueIdentity (the one in `to_uuid`).

This fix prevents these cases to happen checking if `from_id`
matches the `uuid` of its parent UniqueIdentity.

There was some existing tests which have been corrected after
this change and a new test for this fix both for API and Schema.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>