[client:config] Add command to configure the client
The command 'config' has two subcommands that allow
to access the client configuration: 'set' to configure
parameters and 'get' to read the parameters.

By default, the configuration client will be stored
under ".sortinghat/sortinghat.cfg" on the user's
home directory. The parameter '--filename' allows
to define which config file will be used when
these commands are called.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>