[client] Allow to read configuration files
The command will be able to read configuration files
and pass the values to the subcommands.

So far, we only have configuration parameters for the
GraphQL client. For that reason, the parameters are
only handled by the client decorator and not passed
down the chain of commands.

Parameters defined on the client command line will
override any value from the configuration file.
For these reason, default values were removed from
the GraphQL client to determine whether they
were given or not using the command line. With
'click' is not possible to know if the parameter
was provided.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>