[model] Rename UniqueIdentity to Individual
The aim of this commit is to change the terminology
used in SortingHat to name a set of identities.
So far, the name has been 'unique identity' but that
created misunderstanding and confusion in developers
and users of the tool.

The term 'individual' is more clear and easier to
understand than 'unique identity'.

According to the Cambridge dictionary, an individual is
"a single person or thing, especially when compared to
the group or set to which they belong". According to
Merriam-Webster, it is "a particular being or thing as
distinguished from a class, species, or collection".

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>