[model] Update Individual model to add 'mk' (main key) field
The purpose of this change is to use the term unique
identifier or 'uuid' only for identities. Thus, an
individual will be accessed using keys. These keys
are any of its identities UUIDs.

Individuals will still have a main key or reference
to identify them. This field has been called 'mk'
(e.g 'main key').

The field 'uuid' has been removed from the model.
This commit also updates the rest of components to
use the new field and to remove 'uuid' references.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>