[ui] Add UI basic components for testing
To test that the infrastructure created to develop the UI
it is necessary to add some components. This components
are not part of the UI and will be removed as soon as
the first SortingHat UI is designed.

To test the dev infrastructure works, run the commands
`yarn serve` and `yarn storybook` inside the UI dir.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>