[recommendations] Add affiliation recommendations
The function recommend_affiliations() generates a list
of organizations when individuals can be enrolled.
The recommendations are based on the email addreses
of the individuals and the domains of the organizations.

The function checks if the domains of these email
addresses of an individual match with any of the
domains stored on the registry. If this is the case,
the organization associated to that domain will be
recommended.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>