[jobs] Add affiliation job
The affiliation process consist on get which organizations
are valid for an individual (using recommendations) and
enroll it to them. The job will get as an input a list
of individuals identifiers.

As 'affiliation' is a long process it will run in a jobs
queue. The scheduler selected to manage these tasks is 'rq'.
It has an extension for Django ('django-rq') that allows
to visualize the queues, workers and the status of the
jobs. It also needs Redis to handle the communications
between workers and queues.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>