[schema] Add mutation to affiliate individuals
The mutation will schedule a job to affiliate all the
individuals defined by any of the keys passed in `uuids`
argument. When no keys are given, all the individuals
on the registry will be affiliate.

The mutation returns the identifier of the scheduled job.
With this id it will be able to check later the job status
and if its results are ready.

This is an example of the call in GraphQL to
the mutation:

mutation affiliate($uuids: [String]) {
    affiliate(uuids: $uuids) {
        jobId
    }
}

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>