[schema] Add job query
This query retrieves the status of a job. The data
returned contain the id of the job, its status, and
the results of its execution.

Within the data, there is also the field `errors`,
which stores any error raised during the process.

As the data in the results field may be different
depending of the type of job, this field has been
declared as `graphene.Union`. This allows to return
different kind of results. Right now, there is only
one type available: `AffiliationResultType`.

The type `AffiliationResultType` stores the uuid
of the individual and a list with the organizations
it was enrolled to during the affiliation process
run by the job.

This is an example of job query to get the results
generated by an affiliation job:

```
{
  job(
    jobId:"1234-5678-90AB-CDEF"
  ){
    jobId
    status
    errors
    result {
      __typename
      ... on AffiliationResultType {
          uuid
          organizations
      }
    }
  }
}
```

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>