[jobs] Get affiliation recommendations given a set of individuals
This function generates a list of recommendations which include
the organizations where individuals can be affiliated.
This job returns a dictionary with which individuals are
recommended to be affiliated to which organization.

Individuals are defined by any of their valid keys or UUIDs.
When the parameter `uuids` is empty, the job will take all the
individuals stored in the registry.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>