[schema] Add mutation for generate affiliation recommendations
The mutation returns a job which returns a list of
recommendations which include the organizations where individuals
can be affiliated.

If the input list of individuals is empty, the job will be applied
to all the individuals in the registry.

The class `JobType` has been modified to include a new field
called `job_type`, which includes the name of the method that job is
executed for, and its value is obtained from the field `func_name`
inside the `Job` object.

This was needed because the resolver for jobs has to distinguish
between different jobs executed in different methods, as they
can return values in different ways.

Example for the new mutation:

```
mutation {
  recommendAffiliations(uuids: ["<uuid1>", "<uuid2>"]) {
    jobId
  }
}
```

```
{
  job(jobId: "<job_id>"){
    jobId
    status
    jobType
    result {
      __typename
      ... on AffiliationRecommendationType {
          uuid
          organizations
      }
    }
    errors
  }
}
```

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>