[ui] Show individuals from the GraphQL database
This commit modifies the components in order to show, queriying
the GraphQL database, the individuals of the database.

This is the first step, the component App query the database and
save in its data the individuals, then pass the individuals to
the component IndividualsGrid using its props and the
IndividualsGrid component print the name of the "profile" of each
individual.

Signed-off-by: David Moreno <dmoreno@bitergia.com>