[recommendations] Identity matches based on a list of individuals
The function checks if any identity from each individual matches
with a given set of target individuals. First, it filters by the
fields from the criteria and then it groups the matches by
identity. Then, these groups are turned into sets of matching
identities found among all the input identities set.

Each recommendation contains the uuid of the individual provided
in the input list and a list with the matching individuals or
the matching identities if the `verbose` option is activated.

When no matching is found for a given individual, an empty list
will be returned for that individual.

The method returns a generator of identity matches
recommendations.

The algorithm implementing the matching uses `pandas` module,
which has been added to `requirements.txt` file.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>