[jobs] Get matching identities recommendations from sets of individuals
This function generates a list of recommendations which include the
matching identities from the individuals which can be merged with.
This job returns a dictionary with which individuals are recommended
to be merged to which individual (or which identities is `verbose`
mode is activated).

Individuals both for `source_uuids` and `target_uuids` are defined by
any of their valid keys or UUIDs. When the parameter `target_uuids`
is empty, the job will take all the individuals stored in the registry,
so matches will be found comparing the identities from the individuals
in `source_uuids` against all the identities on the registry.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>