[matching] Improve matching recommendations method
When the method receives an empty list of `target_uuids`, the
whole set of identities was retrieved in an inefficient way, as
the former input was a query set with all the individuals on the
registry. Each individual caused two queries, one for retrieving
the individual object and another one for getting its identities.
This was slowing down the performance heavily.

Now, in this case all the identities from the registry are
retrieved with a single, more efficient query.

Before this change, the `target_uuids` were retrieved at `jobs`
level. Now, the `recommend_matches` method is in charge of this
operation.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>